Use `UInt32` for `gitCommitParent()` index parameter
Changed the `n` parameter's type from `UInt` to `UInt32` to remove the
type cast that could potentially overflow on some platforms. Although an
overflow is unlikely in practice, the caller should handle conversions.

+++ Sources/SwiftLibgit2/Commit/Commit-Functions.swift
@@ -484,7 +484,7 @@ public func gitCommitParentCount(
 public func gitCommitParent(
     out     : UnsafeMutablePointer<OpaquePointer?>,
     commit  : OpaquePointer,
-    n       : UInt
+    n       : UInt32
 ) -> GitErrorCode
 {
     return withCConversion
@@ -492,7 +492,7 @@ public func gitCommitParent(
         return git_commit_parent(
             out,
             commit,
-            UInt32(n)
+            n
         )
     }
 }
